Restructure main, split Actions.py, minor clean-up
